+++ src/error_handling/check_binary.c
@@ -36,7 +36,7 @@ int check_binary(char *filename)
 
     if (file == NULL)
         return 84;
-    if (fread(&data, sizeof(header_t), 1, file) != -1) {
+    if (fread(&data, sizeof(header_t), 1, file) != (size_t)-1) {
         fclose(file);
         return 84;
     }
